requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.20.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.7
isodate==0.6.1
kombu==5.3.7
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
pillow==10.4.0
portalocker==2.10.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.9.4
